add firebase logic

+++ src/song/song.service.ts
@@ -7,13 +7,24 @@ import { FirebaseService } from 'src/firebase/firebase.service';
 @Injectable()
 export class SongService {
     constructor(private prisma: PrismaService, private readonly firebaseService: FirebaseService) { }
-    async createSong(albumId: string, dto: SongDto, songPath: string, url: string) {
+    async createSong(albumId: string, dto: SongDto, songFile: Express.Multer.File) {
+        const storage = this.firebaseService.getStorageInstance();
+        const songPath = `song/${albumId}/${songFile.originalname}`;
+        await storage.bucket().upload(songFile.path, {
+            destination: songPath,
+        });
+
+        const song = storage.bucket().file(songPath);
+        const [songUrl] = await song.getSignedUrl({
+            action: 'read',
+            expires: '2099-12-31',
+        });
         return this.prisma.song.create({
             data: {
                 title: dto.title,
                 albumId,
                 songPath,
-                url
+                url: songUrl,
             }
         })
     }
@@ -44,20 +55,31 @@ export class SongService {
             }
         });
         if (!song) {
-            throw new NotFoundDataException('song', songId)
+            throw new HttpException('Songs not found', HttpStatus.NOT_FOUND)
         }
         return song;
     }
-    async updateSong(albumId: string, songId: string, dto: SongDto, songPath: string, url: string) {
+    async updateSong(albumId: string, songId: string, dto: SongDto, songFile: Express.Multer.File) {
         const oldSongData = await this.prisma.song.findUnique({
             where: {
                 id: songId
             }
         })
         if (oldSongData.songPath) {
-            const oldFile = this.firebaseService.getStorageInstance().bucket().file(oldSongData.url);
+            const oldFile = this.firebaseService.getStorageInstance().bucket().file(oldSongData.songPath);
             await oldFile.delete();
         }
+
+        const storage = this.firebaseService.getStorageInstance();
+        const songPath = `song/${albumId}/${songFile.originalname}`;
+        await storage.bucket().upload(songFile.path, {
+            destination: songPath,
+        });
+        const song = storage.bucket().file(songPath);
+        const [songUrl] = await song.getSignedUrl({
+            action: 'read',
+            expires: '2099-12-31',
+        });
         return this.prisma.song.update({
             where: {
                 id: songId
@@ -66,8 +88,7 @@ export class SongService {
                 title: dto.title,
                 albumId,
                 songPath,
-                url
-
+                url: songUrl,
             }
         })
     }
@@ -77,8 +98,10 @@ export class SongService {
                 id: songId
             }
         })
-        const file = this.firebaseService.getStorageInstance().bucket().file(song.url)
-        await file.delete()
+        if (song.songPath) {
+            const file = this.firebaseService.getStorageInstance().bucket().file(song.songPath)
+            await file.delete()
+        }
         return this.prisma.song.delete({
             where: {
                 id: songId
